Initial tree and structure
This is the first commit, so it includes many new files.
The commit includes the basic sturcture expected to remain throughout
the development of the application.

The structure includes a basic Javascript representation of the API.
The API data is being auto-updated using intervals for now, as it is
easier to implement and we expect API changes in this regard.

Signed-off-by: Daniel Zautner <daniel@fewbytes.com>