Added clear methods to views.
In order for the views to be able to render, the views displayed
previously have to be cleared, or else the DOM will become a mess.
To solve this, I have added clear methods to all of the views, which
will be called everytime the route is changed (a new handler is set)

To achive that I have addded a custom event that will be fired on route
changes.