Fix wrong hostname display
This is a fix to issue #13.

The hostname in the top of the html was wrongly displayed. That was due
to hardcoded markup.

To fix the issue, I have added a new page handler, BaseHandler to handle
all of the dynamic features that are out of the scope of specific
handlers (such as the header, which is shared between all of the
handlers).

The base hadnlers listens to the route change event and updates shared
dynamic information.