Removed padding from graphs.
Out of the box jqplot adds padding to the plots on the graph, leaving
space to the right and to the left of it.
This causes some unexpected jumps when the graph updates and a general
feeling that the graph is it to be completely populated.

To solve this the graph box now dynamically set the min and max points
of the x axis based on the data retrieved from the API.

Solves #26