Added check to test if the graph is displayed
When the play event is dispatched, all of the graphs are calling their own play method.
In turn, this method calls the render method, which renders the graph without checking
if it is even displayed at the moment, causing al of the graphs to render over eachother when the
play button is pressed.
This was easily fixed by keeping the displaying state of the graph (should it be displayed?)
as a property for every graph - set to true post render, and to false on clean.