Added proper initalization of parent class to grahs
The graphs in the system all extend the base GraphBox, but they did not
call the constructor of the parent.
As the initalization of the event listeners for the play pause events
happens in the constructor, I have added a proper call to the parent constructor
for all of the classes extending GraphBox.