Fixed dissapearing Threads Timeline
The threads timeline was updated by an interval set in the constructor
of ThreadsLayout.

Because the ThreadLayout class is initialized once, the interval will only
initialize once.
This becomes a problem when the "clear" method is called on route change.
The interval will get cleared, but will not be re set when the threads route
is called again.

Moving the interval setting to the "render" method solved the issue.