Implemented the latest profiler HTML
I have implemented the new profiler HTML and some other small changes.
Between the changes: styled titles for every graph box instead of the
default jqplot ones.

Note that the discussion about the the behaivour of the traces onces selected
is still on going. So this commit does not include any changes to the behaivour
of the tracepoints when picked, and just implment the styling changes.