Fixed API returning empty CPU data
The CPU data tried returning the "data" property while the name
for the property changed to "threads" a while ago.
Given the complete lack of the type saftey in JS, the bug was
fairly hidden, and looked as if there is just no data.

I changed the refernece to "threads" and now the chart and the API are working.