Added strict type check to lastValue
The DerivativePlot#add method was checking for the existence of a previous value.
if there is non, you can calculate the derivative, so you skip it.
When the check was for loose types, and the first value of the plot was 0,
the plot will never render.
Changed the check to a strict one against null and the plots work again.