added BatchRequests class
Added a simple class to abstract the batch requests API.
The class works with a simple queue.
When the 'get' method is called, the request is being added to the queue
and a promise is returned.
There in an invetral that commits all of the requests in a batch once a given time and resets the queue.
When there is a response from the batch api, the requests that were queued are being iterated
and their matching promises are resolved or rejected based on the response from the API.

This allows the older way of sending ajax requests to the server work seemlessly with
the new batch api.