Fixed the main index file build
There was a bug in the build process, where the makefile will call
the build_index script and try to echo the result to the index.html file,
but the script it self already echoed it to the index.html and returned nothing.
So after the index.html file was built by build_index.sh, the makefile will echo an empty string
to it and zero it out.

I moved the actual building of the file to the Makefile and left the build_index.sh script
to only build and return the content.