Added support for priority request in batch api
There was an issue that when deleting a trace from the count,
The delete request will be added to the queue after the get request.
That led to the response to send 'data' even though the trace is no longer there.
To fix the issue I have added an option to add priority request to the begining of the queue.
Assuring that they will execute first.