Fixed division by zero bug
Because jolokia returns the timestamp in seconds,
and we sample twice a second, a divsion by zerion bug occured.
This happend because the derivative in some graphs is calculated like this:
(value - latestValue) / (timestamp - latestTimestamp)
Because the timestamp only changes once a second, timestamp and lastTimestamp
can be equal, leading to divsion by zero and weird jumps in the graph.