OSv GUI: Updating the tab routing
The routes to the tab where missing a slash at the end, in that case when the
GUI was opened on the first time and was redirected to /dashboard/ it did not
point to the main tab.

The same problem would occure when going to a direct link to one of the tabs
(or refreshing).

To solve this, both the routes where modified to match an ending slash and the
href to the tabs was modified.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>